refactor #20 : 리뷰사항 반영
- checkLastPage 메서드를 PaginationRepository로 분리
  - PaginationRepository 생성
- 상품 목록 조회 테스트 코드 수정
  - 기존 : 가져온 사이즈만 체크
  - 변경 : 사이즈 체크 및 객체의 필드가 알맞게 들어왔는지, 생성시간 순으로 정렬되어 있는지 등을 체크.
- ItemPaginationRepository에서 판매자의 닉네임을 isSeller 필드에 바인딩 해줘야하는데 생성되지 않은 필드 sellerId로 잘못가져와서 isSeller필드에는 데이터가 바인딩이 되지 않고 null이 들어옴 -> sellerId -> isSeller로 수정
